Add more tests to ensure 100% coverage
