+++ tests/test_others.py
@@ -208,27 +208,6 @@ def test_autocompletion_too_many_parameters():
     assert exc_info.value.message == "Invalid autocompletion callback parameters: val2"
 
 
-def test_forward_references():
-    app = typer.Typer()
-
-    @app.command()
-    def main(arg1, arg2: int, arg3: "int", arg4: bool = False, arg5: "bool" = False):
-        print(f"arg1: {type(arg1)} {arg1}")
-        print(f"arg2: {type(arg2)} {arg2}")
-        print(f"arg3: {type(arg3)} {arg3}")
-        print(f"arg4: {type(arg4)} {arg4}")
-        print(f"arg5: {type(arg5)} {arg5}")
-
-    result = runner.invoke(app, ["Hello", "2", "invalid"])
-
-    assert "Invalid value for 'ARG3': 'invalid' is not a valid integer" in result.stdout
-    result = runner.invoke(app, ["Hello", "2", "3", "--arg4", "--arg5"])
-    assert (
-        "arg1: <class 'str'> Hello\narg2: <class 'int'> 2\narg3: <class 'int'> 3\narg4: <class 'bool'> True\narg5: <class 'bool'> True\n"
-        in result.stdout
-    )
-
-
 def test_context_settings_inheritance_single_command():
     app = typer.Typer(context_settings={"help_option_names": ["-h", "--help"]})
